Rearrange files and delete obsolete components
