--- packages/plugins/users-permissions/admin/src/components/Permissions/ListWrapper.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-const ListWrapper = styled.div`
-  background-color: ${({ theme }) => theme.main.colors.white};
-  padding-top: 1rem;
-  padding-bottom: 1.8rem;
-`;
-
-export default ListWrapper;
