--- packages/plugins/users-permissions/admin/src/components/Permissions/PermissionRow/BaselineAlignment.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-const BaselineAlignment = styled.div`
-  padding-top: 1px;
-`;
-
-export default BaselineAlignment;
