--- packages/plugins/users-permissions/admin/src/components/Permissions/PermissionRow/SubCategory/ConditionsButtonWrapper.js
@@ -1,13 +1,0 @@
-/* eslint-disable indent */
-import styled from 'styled-components';
-
-const ConditionsButtonWrapper = styled.div`
-  padding: 0.9rem;
-  ${({ hasConditions }) =>
-    hasConditions &&
-    `
-    padding-left: 22px;
-  `}
-`;
-
-export default ConditionsButtonWrapper;
