--- packages/plugins/users-permissions/admin/src/components/Permissions/PermissionRow/SubCategory/PolicyWrapper.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-const PolicyWrapper = styled.div`
-  flex: 1;
-  text-align: end;
-`;
-
-export default PolicyWrapper;
